chore: modify_directory
