+++ src/main/java/dku25/chatGraph/api/graph/repository/QuestionRepository.java
@@ -1,5 +1,6 @@
-package dku25.chatGraph.api.graph;
+package dku25.chatGraph.api.graph.repository;
 
+import dku25.chatGraph.api.graph.node.QuestionNode;
 import org.springframework.data.neo4j.repository.Neo4jRepository;
 
 public interface QuestionRepository extends Neo4jRepository<QuestionNode, Long> {}
